refactor: put dto code into separate files (from interfaces), redo apy.yaml
