+++ package.json
@@ -31,7 +31,6 @@
     "@nestjs/jwt": "^10.2.0",
     "@nestjs/mapped-types": "^2.1.0",
     "@nestjs/platform-express": "^10.4.16",
-    "@nestjs/swagger": "^7.4.2",
     "bcrypt": "^5.1.1",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.1",
