+++ src/albums/album.controller.ts
@@ -9,7 +9,8 @@ import {
   HttpCode,
 } from '@nestjs/common';
 import { AlbumService } from './album.service';
-import { CreateAlbumDto, UpdateAlbumDto } from './interfaces';
+import { CreateAlbumDto } from './dto/create-album.dto';
+import { UpdateAlbumDto } from './dto/update-album.dto';
 
 @Controller('album')
 export class AlbumController {
