+++ src/albums/dto/update-album.dto.ts
@@ -1,0 +1,21 @@
+import {
+  IsString,
+  IsInt,
+  IsNotEmpty,
+  IsOptional,
+  IsUUID,
+} from 'class-validator';
+
+export class UpdateAlbumDto {
+  @IsString()
+  @IsNotEmpty()
+  name: string;
+
+  @IsInt()
+  @IsNotEmpty()
+  year: number;
+
+  @IsOptional()
+  @IsUUID()
+  artistId: string | null;
+}
