+++ src/albums/interfaces.ts
@@ -1,42 +1,6 @@
-import {
-  IsString,
-  IsInt,
-  IsNotEmpty,
-  IsOptional,
-  IsUUID,
-} from 'class-validator';
-
 export class Album {
   id: string;
   name: string;
   year: number;
-  artistId: string | null;
-}
-
-export class CreateAlbumDto {
-  @IsString()
-  @IsNotEmpty()
-  name: string;
-
-  @IsInt()
-  @IsNotEmpty()
-  year: number;
-
-  @IsOptional()
-  @IsUUID()
-  artistId: string | null;
-}
-
-export class UpdateAlbumDto {
-  @IsString()
-  @IsNotEmpty()
-  name: string;
-
-  @IsInt()
-  @IsNotEmpty()
-  year: number;
-
-  @IsOptional()
-  @IsUUID()
   artistId: string | null;
 }
