+++ src/artists/artist.controller.ts
@@ -9,7 +9,8 @@ import {
   HttpCode,
 } from '@nestjs/common';
 import { ArtistService } from './artist.service';
-import { CreateArtistDto, UpdateArtistDto } from './interfaces';
+import { CreateArtistDto } from './dto/creat-artist.dto';
+import { UpdateArtistDto } from './dto/update-artist.dto';
 
 @Controller('artist')
 export class ArtistController {
