+++ src/artists/dto/creat-artist.dto.ts
@@ -1,0 +1,10 @@
+import { IsString, IsBoolean, IsNotEmpty } from 'class-validator';
+
+export class CreateArtistDto {
+  @IsString()
+  @IsNotEmpty()
+  name: string;
+
+  @IsBoolean()
+  grammy: boolean;
+}
