+++ src/main.ts
@@ -1,25 +1,14 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import { ConfigService } from '@nestjs/config';
-import { HttpExceptionFilter } from './filters/http-exception.filter';
-import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
+import { ValidationPipe } from '@nestjs/common';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
   const configService = app.get(ConfigService);
   const port = configService.get<number>('PORT') || 4000;
 
-  app.useGlobalFilters(new HttpExceptionFilter());
-
-  const config = new DocumentBuilder()
-    .setTitle('Home Library API')
-    .setDescription(
-      'API for managing users, artists, tracks, albums, and favorites',
-    )
-    .setVersion('1.0')
-    .build();
-  const document = SwaggerModule.createDocument(app, config);
-  SwaggerModule.setup('api', app, document);
+  app.useGlobalPipes(new ValidationPipe({ transform: true }));
 
   await app.listen(port);
   console.log(`Server is running on port ${port}`);
