+++ src/tracks/dto/update-track.dto.ts
@@ -1,0 +1,25 @@
+import {
+  IsString,
+  IsInt,
+  IsNotEmpty,
+  IsOptional,
+  IsUUID,
+} from 'class-validator';
+
+export class UpdateTrackDto {
+  @IsString()
+  @IsNotEmpty()
+  name: string;
+
+  @IsOptional()
+  @IsUUID()
+  artistId: string | null;
+
+  @IsOptional()
+  @IsUUID()
+  albumId: string | null;
+
+  @IsInt()
+  @IsNotEmpty()
+  duration: number;
+}
