+++ src/tracks/interfaces.ts
@@ -1,51 +1,7 @@
-import {
-  IsString,
-  IsInt,
-  IsNotEmpty,
-  IsOptional,
-  IsUUID,
-} from 'class-validator';
-
 export class Track {
   id: string;
   name: string;
   artistId: string | null;
   albumId: string | null;
-  duration: number;
-}
-
-export class CreateTrackDto {
-  @IsString()
-  @IsNotEmpty()
-  name: string;
-
-  @IsOptional()
-  @IsUUID()
-  artistId: string | null;
-
-  @IsOptional()
-  @IsUUID()
-  albumId: string | null;
-
-  @IsInt()
-  @IsNotEmpty()
-  duration: number;
-}
-
-export class UpdateTrackDto {
-  @IsString()
-  @IsNotEmpty()
-  name: string;
-
-  @IsOptional()
-  @IsUUID()
-  artistId: string | null;
-
-  @IsOptional()
-  @IsUUID()
-  albumId: string | null;
-
-  @IsInt()
-  @IsNotEmpty()
   duration: number;
 }
