+++ src/tracks/track.controller.ts
@@ -9,7 +9,8 @@ import {
   HttpCode,
 } from '@nestjs/common';
 import { TrackService } from './track.service';
-import { CreateTrackDto, UpdateTrackDto } from './interfaces';
+import { CreateTrackDto } from './dto/create-track.dto';
+import { UpdateTrackDto } from './dto/update-track.dto';
 
 @Controller('track')
 export class TrackController {
